+++ htdocs/modules/handler.php
@@ -5,14 +5,23 @@ class regexException extends Exception {}
 class databaseConnectException extends Exception {}
 // ENVIRONMENT NOT FOUND ERROR
 class environmentNotFoundException extends Exception {}
-// ENVIRONMENT VARIABLE NOT FOUND ERROR
-class environmentVariableNotFoundException extends Exception {}
 // MODULE NOT FOUND ERROR
 class moduleNotFoundException extends Exception {}
+// TOO MANY ARGUMENTS ERROR
+class tooManyArgumentsException extends Exception {}
+// NOT ENOUGH ARGUMENTS ERROR
+class notEnoughArgumentsException extends Exception {}
+// UNKNOWN ARGUMENT VALUE
+class unknownArgumentValueException extends Exception {}
 
 // EXCEPTION HANDLER
 function exceptionHandler(Throwable $exception) {
-    echo "Error ({$exception->getFile()}:{$exception->getLine()}:{" . get_class($exception) . "}) -> {$exception->getMessage()}";
+    echo json_encode([
+        "Error" => get_class($exception), 
+        "File" => $exception->getFile(), 
+        "Line" => $exception->getLine(),
+        "Message" => $exception->getMessage()
+    ]);
     exit;
 }
 set_exception_handler("exceptionHandler");